Implemented an abstract two-way queue communication protocol.
ConvergenceService waits for answers by using synchronization.
The answer is along the lines of OK or error message as specified.
Created a full abstraction QueueAccess library with opt-in features.
The request id is being utilized to mark messages. It's even hashed.
ConvergenceService answers properly assign queue name postfix.
After some troubles, it was decided that postfix within the message
mustn't be null; though it can be an empty string. It is implemented
in a lazy way, so that the null only crashes after it's requested!
Added some constructor not-null assertion, mainly for Configuration.
Migrating convergence controller methods to use async Task,
because the previous implementation displayed exceptions caught
via Wait() very poorly, even when the InnerException was used!